requirements.txt
psycopg2
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
sqlalchemy-utils
flask_login